To get to a workable format: simply edit "drawControlPoints" to only take the first argument ("selectedCurve") and uncomment "selectedControlPoint" in Freeform.h.
Selecting Control Points:

The issue is that I have defined a variable selectedControlPoint within each curve but coded all of main as if it were a global variable. I can't switch it to a global variable because then it would be outside of the scope of Freeform and then I'd have to pass it into the draw function every time.

Options:
    pass selectedControlPoint into "drawControlPoints" and "deleteControlPoints" and maybe other things. (this would also mean that a curve that is not selected cannot have any selected control points - which I'm not sure is good or bad)
    deal with having a selectedControlPoint variable in each curve.